+++ gridsome/lib/app/queue/ImageProcessQueue.js
@@ -2,13 +2,14 @@ const path = require('path')
 const fs = require('fs-extra')
 const sharp = require('sharp')
 const crypto = require('crypto')
-const { trim } = require('lodash')
 const mime = require('mime-types')
 const colorString = require('color-string')
 const md5File = require('md5-file/promise')
 const imageSize = require('probe-image-size')
 const svgDataUri = require('mini-svg-data-uri')
 const { forwardSlash } = require('../../utils')
+const { warmupSharp } = require('../../utils/sharp')
+const { reject } = require('lodash')
 
 class ImageProcessQueue {
   constructor ({ context, config }) {
@@ -22,27 +23,23 @@ class ImageProcessQueue {
   }
 
   async add (filePath, options = {}) {
-    let asset
-
-    try {
-      asset = await this.preProcess(filePath, options)
-    } catch (err) {
-      throw err
-    }
+    const asset = await this.preProcess(filePath, options)
 
     if (process.env.GRIDSOME_MODE === 'serve') {
       return asset
     }
 
-    asset.sets.forEach(({ filename, src, width }) => {
-      if (!this._queue.has(src + asset.cacheKey)) {
-        this._queue.set(src + asset.cacheKey, {
-          destination: trim(src, this.config.pathPrefix),
-          options: { ...options, width },
+    asset.sets.forEach(({ filename, destPath, width, height }) => {
+      if (!this._queue.has(destPath + asset.cacheKey)) {
+        this._queue.set(destPath + asset.cacheKey, {
+          options: { ...options, width, height },
           cacheKey: asset.cacheKey,
           size: asset.size,
+          destPath,
           filename,
-          filePath
+          filePath,
+          width: asset.width,
+          height: asset.height
         })
       }
     })
@@ -51,8 +48,9 @@ class ImageProcessQueue {
   }
 
   async preProcess (filePath, options = {}) {
-    const { imageExtensions, outDir, pathPrefix, maxImageWidth } = this.config
-    const imagesDir = path.relative(outDir, this.config.imagesDir)
+    const { imageExtensions, outputDir, pathPrefix, maxImageWidth } = this.config
+    const { minSizeDistance = 300 } = this.config.images || {}
+    const imagesDir = path.relative(outputDir, this.config.imagesDir)
     const relPath = path.relative(this.context, filePath)
     const { name, ext } = path.parse(filePath)
     const mimeType = mime.lookup(filePath)
@@ -70,29 +68,27 @@ class ImageProcessQueue {
 
     const hash = await md5File(filePath)
     const buffer = await fs.readFile(filePath)
-    const { width, height } = imageSize.sync(buffer)
+    const originalSize = imageSize.sync(buffer) || { width: 1, height: 1 }
 
-    const imageWidth = Math.min(
-      parseInt(options.width || width, 10),
-      maxImageWidth,
-      width
-    )
+    const { imageWidth, imageHeight } = computeScaledImageSize(originalSize, options, maxImageWidth)
 
-    const imageHeight = options.height !== undefined
-      ? parseInt(options.height, 10)
-      : Math.ceil(height * (imageWidth / width))
+    let imageWidths = options.imageWidths || [480, 1024, 1920, 2560]
 
-    const imageSizes = (options.sizes || [480, 1024, 1920, 2560]).filter(size => {
-      return size <= maxImageWidth && size <= imageWidth
-    })
+    if (typeof imageWidths === 'string') {
+      imageWidths = imageWidths.split(',')
+    }
 
-    if (
-      (imageSizes.length === 1 && imageSizes[0] <= imageWidth) ||
-      (imageSizes.length === 0)
-    ) {
-      if (imageWidth <= maxImageWidth) {
+    let imageSizes = imageWidths.filter(size => size <= imageWidth)
+    const maxWidth = Math.max(...imageSizes, 0)
+
+    if (!options.imageWidths) {
+      if (imageWidth > maxWidth || imageSizes.length === 0) {
         imageSizes.push(imageWidth)
       }
+
+      imageSizes = reject(imageSizes, (width, i, arr) => {
+        return arr[i + 1] - width < minSizeDistance
+      })
     }
 
     // validate color string
@@ -102,36 +98,42 @@ class ImageProcessQueue {
       options.background = this.config.imageBackgroundColor
     }
 
-    const createSrcPath = (filename, imageOptions) => {
+    const cacheKey = genHash(filePath + hash + JSON.stringify(options)).substr(0, 7)
+
+    const createDestPath = (filename, imageOptions) => {
       if (process.env.GRIDSOME_MODE === 'serve') {
-        const query = '?' + createOptionsQuery(imageOptions)
+        const key = process.env.GRIDSOME_TEST ? 'test' : cacheKey
+        const query = '?' + createOptionsQuery(imageOptions.concat({ key: 'key', value: key }))
         return path.join('/', imagesDir, forwardSlash(relPath)) + query
       }
 
-      return forwardSlash(path.join('/', pathPrefix, imagesDir, filename))
+      return path.join(imagesDir, filename)
     }
 
-    const sets = imageSizes.map((width = imageWidth) => {
-      const height = Math.ceil(imageHeight * (width / imageWidth))
-      const imageOptions = { ...options, width }
+    const sets = imageSizes.map(width => {
+      let height
 
-      if (options.height !== undefined) {
-        imageOptions.height = height
+      if (options.height) {
+        height = Math.ceil(imageHeight * (width / imageWidth))
       }
 
-      const arr = this.createImageOptions(imageOptions)
+      const arr = this.createImageOptions({ ...options, width, height })
       const filename = this.createFileName(filePath, arr, hash)
-      const src = createSrcPath(filename, arr)
+      const relPath = createDestPath(filename, arr)
+      const destPath = path.join(this.config.outputDir, relPath)
+      const src = encodeURI(forwardSlash(path.join(pathPrefix || '/', relPath)))
 
-      return { filename, src, width, height }
+      return { filename, destPath, src, width, height }
     })
 
     const results = {
       src: sets[sets.length - 1].src,
       size: { width: imageWidth, height: imageHeight },
-      cacheKey: genHash(filePath + hash + JSON.stringify(options)),
+      width: originalSize.width,
+      height: originalSize.height,
       noscriptHTML: '',
       imageHTML: '',
+      cacheKey,
       name,
       ext,
       hash,
@@ -143,7 +145,11 @@ class ImageProcessQueue {
     const isLazy = options.immediate === undefined
 
     if (isSrcset) {
-      results.dataUri = await createDataUri(buffer, mimeType, imageWidth, imageHeight, options)
+      try {
+        results.dataUri = await createDataUri(buffer, mimeType, imageWidth, imageHeight, options)
+      } catch (err) {
+        throw new Error(`Failed to process image ${relPath}. ${err.message}`)
+      }
       results.sizes = options.sizes || `(max-width: ${imageWidth}px) 100vw, ${imageWidth}px`
       results.srcset = results.sets.map(({ src, width }) => `${src} ${width}w`)
     }
@@ -202,6 +208,10 @@ class ImageProcessQueue {
       imageOptions.push({ key: 'background', shortKey: 'b-', value: options.background })
     }
 
+    if (options.blur) {
+      imageOptions.push({ key: 'blur', shortKey: 'bl-', value: options.blur })
+    }
+
     return imageOptions
   }
 
@@ -210,12 +220,63 @@ class ImageProcessQueue {
     const string = arr.length ? createOptionsQuery(arr) : ''
 
     const optionsHash = genHash(string).substr(0, 7)
-    const contentHash = !process.env.GRIDSOME_TEST
-      ? hash.substr(0, 7)
-      : 'test'
+    const contentHash = !process.env.GRIDSOME_TEST ? hash : 'test'
 
     return `${name}.${optionsHash}.${contentHash}${ext}`
   }
+}
+
+function computeScaledImageSize (originalSize, options, maxImageWidth) {
+  const { width, height, fit = 'cover' } = options
+
+  const targetWidth = width || originalSize.width
+  const targetHeight = height || originalSize.height
+
+  if (width && height && ['cover', 'fill', 'contain'].includes(fit)) {
+    return {
+      imageWidth: targetWidth,
+      imageHeight: targetHeight
+    }
+  }
+
+  // Special handling for fit inside and fit outside to prevent blurry images
+  // and page reflow when the images are loaded lazily.
+  // Calculates the scaled size of the image according to the equations found in
+  // https://github.com/lovell/sharp/blob/master/src/pipeline.cc (see MIN and MAX)
+  if (['inside', 'outside'].includes(fit)) {
+    const xFactor = originalSize.width / targetWidth
+    const yFactor = originalSize.height / targetHeight
+
+    let imageWidth = targetWidth
+    let imageHeight = targetHeight
+
+    if (
+      (fit === 'inside' && xFactor > yFactor) ||
+      (fit === 'outside' && xFactor < yFactor)
+    ) {
+      imageHeight = Math.round(originalSize.height / xFactor)
+    } else {
+      imageWidth = Math.round(originalSize.width / yFactor)
+    }
+
+    return { imageWidth, imageHeight }
+  }
+
+  const imageWidth = Math.min(
+    width || originalSize.width,
+    maxImageWidth,
+    originalSize.width
+  )
+
+  let imageHeight = height !== undefined
+    ? height
+    : Math.ceil(originalSize.height * (imageWidth / originalSize.width))
+
+  if (height && height > originalSize.height) {
+    imageHeight = originalSize.height
+  }
+
+  return { imageWidth, imageHeight }
 }
 
 ImageProcessQueue.uid = 0
@@ -231,17 +292,19 @@ function createOptionsQuery (arr) {
 }
 
 async function createDataUri (buffer, type, width, height, options = {}) {
-  const blur = options.blur !== undefined ? parseInt(options.blur, 10) : 40
+  const blur = options.blur !== undefined ? options.blur : 40
   const resizeOptions = {}
 
   if (options.fit) resizeOptions.fit = sharp.fit[options.fit]
   if (options.position) resizeOptions.position = sharp.position[options.position]
   if (options.background) resizeOptions.background = options.background
 
+  const blurredSvg = await createBlurSvg(buffer, type, width, height, blur, resizeOptions)
+
   return svgDataUri(
     `<svg fill="none" viewBox="0 0 ${width} ${height}" ` +
     `xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">` +
-    (blur > 0 ? await createBlurSvg(buffer, type, width, height, blur, resizeOptions) : '') +
+    blurredSvg +
     `</svg>`
   )
 }
@@ -249,17 +312,28 @@ async function createDataUri (buffer, type, width, height, options = {}) {
 async function createBlurSvg (buffer, mimeType, width, height, blur, resize = {}) {
   const blurWidth = 64
   const blurHeight = Math.round(height * (blurWidth / width))
-  const blurBuffer = await sharp(buffer).resize(blurWidth, blurHeight, resize).toBuffer()
+  const warmSharp = await warmupSharp(sharp)
+  const blurBuffer = await warmSharp(buffer).resize(blurWidth, blurHeight, resize).toBuffer()
   const base64 = blurBuffer.toString('base64')
   const id = `__svg-blur-${ImageProcessQueue.uid++}`
+  let defs = ''
 
-  return '' +
-    '<defs>' +
-    `<filter id="${id}">` +
-    `<feGaussianBlur in="SourceGraphic" stdDeviation="${blur}"/>` +
-    `</filter>` +
-    '</defs>' +
-    `<image x="0" y="0" filter="url(#${id})" width="${width}" height="${height}" xlink:href="data:${mimeType};base64,${base64}" />`
+  if (blur > 0) {
+    defs = '' +
+      '<defs>' +
+      `<filter id="${id}">` +
+      `<feGaussianBlur in="SourceGraphic" stdDeviation="${blur}"/>` +
+      `</filter>` +
+      '</defs>'
+  }
+
+  const image = '' +
+    `<image x="0" y="0" ` +
+    (defs ? `filter="url(#${id})" ` : ' ') +
+    `width="${width}" height="${height}" ` +
+    `xlink:href="data:${mimeType};base64,${base64}" />`
+
+  return defs + image
 }
 
 // async function createTracedSvg (buffer, type, width, height) {