+++ gridsome/lib/webpack/modules/assets.js
@@ -1,6 +1,7 @@
 const isUrl = require('is-url')
-const crypto = require('crypto')
+const camelcase = require('camelcase')
 const isRelative = require('is-relative')
+const { isMailtoLink, isTelLink } = require('../../utils')
 
 module.exports = () => ({
   postTransformNode (node) {
@@ -9,25 +10,14 @@ module.exports = () => ({
     }
 
     if (node.tag === 'g-image') {
-      if (!node.hasOwnProperty('key')) {
-        // Enforce a unique key to prevent the g-image component from
-        // re-rendering each time the parent is updated. Or else the
-        // IntersectionObserver will trigger a lazy load on each update.
-
-        const id = nodePath(node)
-        const hash = crypto.createHash('md5')
-        const src = getAttrValue(node, 'src') || ''
-        const key = hash.update(src + id).digest('hex').substr(0, 7)
-
-        node.key = JSON.stringify(`g-image-${key}`)
-      }
-
       transformNodeAttr(node, 'src')
     }
   }
 })
 
 function transformNodeAttr (node, attrName) {
+  if (!Array.isArray(node.attrs)) return
+
   for (const attr of node.attrs) {
     if (attr.name === attrName) {
       if (isStatic(attr.value)) {
@@ -42,7 +32,7 @@ function transformAttrValue (node, attr) {
   const value = extractValue(attr.value)
   let result = attr.value
 
-  if (!isUrl(value) && isRelative(value)) {
+  if (!isUrl(value) && !isMailtoLink(value) && !isTelLink(value) && isRelative(value)) {
     const query = createOptionsQuery(node.attrs)
     result = `require("!!assets-loader?${query}!${value}")`
   }
@@ -51,31 +41,18 @@ function transformAttrValue (node, attr) {
 }
 
 function isStatic (value) {
-  return /^\"[^"]+\"$/.test(value)
+  return /^"[^"]+"$/.test(value)
 }
 
 function extractValue (value) {
   return value.substr(1, value.length - 2)
 }
 
-function getAttrValue (node, attrName) {
-  return node.attrs
-    .filter(({ name }) => name === attrName)
-    .map(({ value }) => isStatic(value) ? extractValue(value) : value)
-    .pop()
-}
-
 function createOptionsQuery (attrs) {
   return attrs
     .filter(attr => attr.name !== 'src')
     .filter(attr => isStatic(attr.value))
-    .map(attr => ({ name: attr.name, value: extractValue(attr.value) }))
+    .map(attr => ({ name: camelcase(attr.name), value: extractValue(attr.value) }))
     .map(attr => `${attr.name}=${encodeURIComponent(attr.value)}`)
     .join('&')
-}
-
-function nodePath (node) {
-  return node && node.parent && Array.isArray(node.parent.children)
-    ? `${nodePath(node.parent)}-${node.parent.children.indexOf(node)}`
-    : ''
 }