+++ gridsome/lib/workers/html-writer.js
@@ -2,31 +2,40 @@ const fs = require('fs-extra')
 const createRenderFn = require('../server/createRenderFn')
 
 exports.render = async function ({
+  hash,
   pages,
   htmlTemplate,
   clientManifestPath,
-  serverBundlePath
+  serverBundlePath,
+  prefetch,
+  preload
 }) {
+  const regexpPrefetch = (prefetch && (typeof(prefetch.mask) === 'string')) ? new RegExp(prefetch.mask) : null
+  const regexpPreload = (preload && (typeof(preload.mask) === 'string')) ? new RegExp(preload.mask) : null
   const render = createRenderFn({
     htmlTemplate,
     clientManifestPath,
-    serverBundlePath
+    serverBundlePath,
+    shouldPrefetch: regexpPrefetch ? file => regexpPrefetch.test(file) : null,
+    shouldPreload: regexpPreload ? file => regexpPreload.test(file) : null
   })
 
-  let page, html
+  let page, html, state, stateSize
+  const length = pages.length
 
-  for (let i = 0, l = pages.length; i < l; i++) {
+  for (let i = 0; i < length; i++) {
     page = pages[i]
+    state = undefined
+    stateSize = undefined
 
-    const { data } = page.dataOutput
-      ? await fs.readJson(page.dataOutput)
-      : {}
+    if (page.dataOutput) {
+      const content = await fs.readFile(page.dataOutput, 'utf8')
 
-    try {
-      html = await render(page.path, data)
-    } catch (err) {
-      throw err
+      stateSize = content.length
+      state = JSON.parse(content)
     }
+
+    html = await render(page, state, stateSize, hash)
 
     await fs.outputFile(page.htmlOutput, html)
   }