+++ gridsome/lib/workers/image-processor.js
@@ -1,22 +1,22 @@
 const path = require('path')
 const fs = require('fs-extra')
+const pMap = require('p-map')
 const sharp = require('sharp')
 const imagemin = require('imagemin')
 const colorString = require('color-string')
 const imageminWebp = require('imagemin-webp')
+const imageminMozjpeg = require('imagemin-mozjpeg')
 const imageminPngquant = require('imagemin-pngquant')
-const imageminJpegoptim = require('imagemin-jpegoptim')
-
-sharp.simd(true)
+const sysinfo = require('../utils/sysinfo')
+const { warmupSharp } = require('../utils/sharp')
 
 exports.processImage = async function ({
+  width,
+  height,
   filePath,
   destPath,
   cachePath,
-  size,
   options = {},
-  minWidth = 500,
-  resizeImage = false,
   backgroundColor = null
 }) {
   if (cachePath && await fs.exists(cachePath)) {
@@ -30,19 +30,22 @@ exports.processImage = async function ({
     const config = {
       pngCompressionLevel: parseInt(options.pngCompressionLevel, 10) || 9,
       quality: parseInt(options.quality, 10) || 75,
-      width: parseInt(options.width, 10),
-      height: parseInt(options.height, 10),
+      width: parseInt(options.width, 10) || null,
+      height: parseInt(options.height, 10) || null,
       jpegProgressive: true
     }
 
     const plugins = []
     let pipeline = sharp(buffer)
 
-    if (config.width && config.width <= size.width) {
-      const ratio = size.height / size.width
-      const height = Math.round(config.width * ratio)
+    if (
+      (config.width && config.width <= width) ||
+      (config.height && config.height <= height)
+    ) {
       const resizeOptions = {}
 
+      if (config.height) resizeOptions.height = config.height
+      if (config.width) resizeOptions.width = config.width
       if (options.fit) resizeOptions.fit = sharp.fit[options.fit]
       if (options.position) resizeOptions.position = sharp.position[options.position]
       if (options.background && colorString.get(options.background)) {
@@ -51,16 +54,18 @@ exports.processImage = async function ({
         resizeOptions.background = backgroundColor
       }
 
-      pipeline = pipeline.resize(config.width, height, resizeOptions)
+      pipeline = pipeline.resize(resizeOptions)
     }
 
     if (/\.png$/.test(ext)) {
+      const quality = config.quality / 100
+
       pipeline = pipeline.png({
         compressionLevel: config.pngCompressionLevel,
         adaptiveFiltering: false
       })
       plugins.push(imageminPngquant({
-        quality: config.quality
+        quality: [quality, quality]
       }))
     }
 
@@ -69,8 +74,9 @@ exports.processImage = async function ({
         progressive: config.jpegProgressive,
         quality: config.quality
       })
-      plugins.push(imageminJpegoptim({
-        max: config.quality
+      plugins.push(imageminMozjpeg({
+        progressive: config.jpegProgressive,
+        quality: config.quality
       }))
     }
 
@@ -90,17 +96,23 @@ exports.processImage = async function ({
   await fs.outputFile(destPath, buffer)
 }
 
-exports.process = async function ({ queue, outDir, cacheDir, minWidth, backgroundColor }) {
-  return Promise.all(queue.map(set => {
+exports.process = async function ({ queue, context, cacheDir, backgroundColor }) {
+  await warmupSharp(sharp)
+  await pMap(queue, async set => {
     const cachePath = cacheDir ? path.join(cacheDir, set.filename) : null
-    const destPath = path.join(outDir, set.destination)
 
-    return exports.processImage({
-      backgroundColor,
-      cachePath,
-      destPath,
-      minWidth,
-      ...set
-    })
-  }))
+    try {
+      await exports.processImage({
+        destPath: set.destPath,
+        backgroundColor,
+        cachePath,
+        ...set
+      })
+    } catch (err) {
+      const relPath = path.relative(context, set.filePath)
+      throw new Error(`Failed to process image ${relPath}. ${err.message}`)
+    }
+  }, {
+    concurrency: sysinfo.cpus.logical
+  })
 }